supprimer tag back

+++ app/deleteTag.php
@@ -2,19 +2,11 @@
 
 header('content-type:application/json');
 
-
-
 require('app/class/tag.class.php');
-
-
-
-
 
 $path = preg_replace('/wp-content(?!.*wp-content).*/','',__DIR__);
 
 include($path.'wp-load.php');
-
-
 
 if(!checkAuthorized(true)){
 
@@ -22,31 +14,23 @@ if(!checkAuthorized(true)){
 
 }
 
-$id = $_GET["idTag"];
+$id = $_GET["id"];
 
 $table = "tag";
 
-$queryMod = $wpdb->get_results("SELECT id FROM module where tag_id = ".$id." ");
+// $query = $wpdb->get_var("SELECT count(quiz.id) as nbUse FROM `tag`LEFT JOIN quiz ON quiz.tag_id = tag.id LEFT JOIN module ON module.tag_id = tag.id  WHERE tag.id =".$id." GROUP BY tag.name");
 
-$queryQ =  $wpdb->get_results("SELECT id FROM quiz where tag_id = ".$id." ");
+$queryMod = $wpdb->get_var("SELECT count(id) FROM module where tag_id = ".$id." ");
 
-if(){
+$queryQ =  $wpdb->get_var("SELECT count(id) FROM quiz where tag_id = ".$id." ");
+
+if($queryMod == 0 && $queryQ == 0){
 
     $wpdb->delete($table, array('id' => $id));
 
 }
 
-
-
-
-
-
-
-
-
-
-
-
+wp_redirect( home_url()."/ajouter-tag" ); 
 
 ?>
